CMFCPropertyGridProperty::TextToVar will convert int var to long var

+++ ImgRegion/ImgRegionDoc.cpp
@@ -70,9 +70,9 @@ void CImgRegion::CreateProperties(CPropertiesWnd * pPropertiesWnd)
 	pLocal->AddSubItem(pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectL] = pProp);
 	pProp = new CSimpleProp(_T("top"), (_variant_t)m_Rect.top, _T("top"), CPropertiesWnd::PropertyItemRectT);
 	pLocal->AddSubItem(pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectT] = pProp);
-	pProp = new CSimpleProp(_T("Width"), (_variant_t)m_Rect.Width(), _T("Width"), CPropertiesWnd::PropertyItemRectW);
+	pProp = new CSimpleProp(_T("Width"), (_variant_t)(LONG)m_Rect.Width(), _T("Width"), CPropertiesWnd::PropertyItemRectW);
 	pLocal->AddSubItem(pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectW] = pProp);
-	pProp = new CSimpleProp(_T("Height"), (_variant_t)m_Rect.Height(), _T("Height"), CPropertiesWnd::PropertyItemRectH);
+	pProp = new CSimpleProp(_T("Height"), (_variant_t)(LONG)m_Rect.Height(), _T("Height"), CPropertiesWnd::PropertyItemRectH);
 	pLocal->AddSubItem(pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectH] = pProp);
 	pGroup->AddSubItem(pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRect] = pLocal);
 
@@ -170,8 +170,8 @@ void CImgRegion::UpdateProperties(CPropertiesWnd * pPropertiesWnd)
 	((CFileProp *)pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemImage])->SetValue((_variant_t)m_ImageStr);
 	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectL]->SetValue((_variant_t)m_Rect.left);
 	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectT]->SetValue((_variant_t)m_Rect.top);
-	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectW]->SetValue((_variant_t)m_Rect.Width());
-	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectH]->SetValue((_variant_t)m_Rect.Height());
+	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectW]->SetValue((_variant_t)(LONG)m_Rect.Width());
+	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemRectH]->SetValue((_variant_t)(LONG)m_Rect.Height());
 	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemBorderX]->SetValue((_variant_t)m_Border.x);
 	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemBorderY]->SetValue((_variant_t)m_Border.y);
 	pPropertiesWnd->m_pProp[CPropertiesWnd::PropertyItemBorderZ]->SetValue((_variant_t)m_Border.z);
